+++ src/config.ts
@@ -3,8 +3,25 @@ export const CONFIG = {
 		NAME: "My Project",
 		VERSION: "1.0.0",
 		DESCRIPTION: "My Project Description",
-		AUTHOR: "My Name",
-		HOMEPAGE: "https://myhomepage.com",
+		AUTHORS: [
+			{
+				url: "https://myhomepage.com",
+				name: "My Name",
+			},
+		],
+		DEVELOPERS: ["Codixfy", "Jonathan Fernandez"],
+		PUBLISHER: "Codixfy",
+		HOMEPAGE: "http://localhost:3000",
+		KEYWORDS: ["My", "Project", "Keywords"],
+		ALTERNATES: {
+			canonical: "http://localhost:3000",
+		},
+		VERIFICATIONS: {
+			google: "",
+			yandex: "",
+			me: "",
+			yahoo: "",
+		},
 	},
 	API: {
 		ROOT: "http://localhost:3000",