+++ vitest.config.ts
@@ -9,7 +9,7 @@ export default defineConfig({
         globals: true,
         environment: "jsdom",
         coverage: {
-            exclude: ["node_modules", ".next/*", "*.mjs", "*.config.*", "src/**/*.bench.*", "src/**/*.test.*", "src/app/layout.tsx"],
+            exclude: ["node_modules", ".next/*", "*.mjs", "*.config.*", "src/**/*.bench.*", "src/**/*.test.*", "src/app/layout.tsx", "**/SEO.ts"],
             thresholds: {
                 lines: 100,
                 functions: 100,